Fix typo in SendAsync, add ReplyAsync and message-based overloads
In most use cases, we'd have a Message from the user at hand. Therefore, it significantly simplifies things if we can just Reply/React/MarkRead/Send back to the sender of the message, so we add extension method overloads.

Also, since attempting to use SendAsync(number, "hello") is an obvious mistake that's too easy to make (the payload should be an object matching expected payloads from WhatsApp API), we add an analyzer that flags this as a compilation error so it can never happen. Ideally, we'd like to exclude a string from being acceptable as an `object`, but this is the next best thing :).