JQ dependency should exclude contentFiles
Since we compile the JQ.cs type in our package itself, we don't need to make that a transitive dep for consumers of our package.

If they want to access the JQ type themselves, they'd need to install JQ explicitly. This is more aligned with the intended usage of the JQ package.