Generalize the former reengage as a general solution
WhatsApp for Business doesn't have the concept of reengaging or anything like it. It's just templates, with or without complex parameters.

Since we don't offer a full typed model for template objects (for now?) we just let the user construct the object themselves and we just provide the simple template as a ctor overload on the TemplateResponse instead.

This is a balance between simplicy and a comprehensive API for template messages that is perhaps part of another feature.