Make usability extension methods public
We should keep the public API to send messages directly through the IWhatsAppClient. This makes the app more reusable in scenarios that aren't implemented directly as functionality pipelines (i.e. async or scheduled jobs that notify users of updates).

The functionality pipeline approach works best for request/reponse scenarios where the user is driving a conversation, but it's not the only scenario we want to cover with this library.