Improve feature filter naming to avoid potential collisions
A filter simply named "FeatureFilter" as an alias might collide with consumers that also use filters for their own functionality.

So we make it more unique.

Also, since leveraging feature flags is something we use as an internal implementation detail, we make the FeatureFlags enum internal instead, since it's not intended for public consumption.