Disable inherntly flaky media resolving test
In order for the media resolve to work, an existing message that is never deleted needs to exist in the referenced whatsapp number. This is highly brittle and therefore doesn't add much value.

The test can be un-skipped with an actual message/media combination should re-testing be necessary at some point.