Revamp and simplify conversation management
We had a complicated mechanism that relied in feature flags based on presence of services, split storage between message and conversation, forced sequencing outside of the pipeline configuration, etc.

This revamp achieves the following:

- Single handler and extension method to configure conversations `UseConversation`.
- Single storage manager that uses a two-strategy approach:
   * Table storage for active conversation header and individual messages
   * Blob storage for full conversations (due to size restrictions on table storage that may be hit).
- Honoring the user's preference in the pipeline where the conversation loading/restoring happens: we're now just a regular handler like any other.

We also make the target storage account pluggable via DI, but with a sensible OOB default that leverages the built-in AzureWebJobsStorage setting which is already being used by the queue and table for deduplication.