Allow pipeline handlers to send messages too
The previous approach assumed responses would only be generated by the target handler, not the intermediate pipeline. This meant that reactions or other cross-cutting messages sent by other stages in the pipeline would be missed by the sender, since it wrapped only around the target handler in the pipeline.

We now wrap around the outermost handler too, and the sending handler (which will have two instances running now) only sends unsent messages it sees (since the innermost will have sent the ones generated by the target).

Obviously, this means that the placement of the UseConversation handler is relevant since it persists responses which have a message ID, which will only be the case for messages that reach it *after* the target and its innermost sender. Pipeline-generated messages that run before the conversation one will therefore not be saved (as expected) but will still be sent out by the outermost handler.