Make sure we don't lose config in CLI
Since the dotnet tool is run from a shim in a folder different from the one we had the appsettings.json in, we were polluting the output with logs.

By pulling those settings from memory, we ensure they are always present. Yet they are still overridable from secrets or dotnetconfig.