Massive timeout increase to aid local debugging
Non-timeouts should be the default since this is a local development tool aid, and while debugging your local WhatsApp functionality pipelines, it can take quite a while for the response to come back. We don't want the client aborting the request to the server and causing an unexpected azure functions runtime error due to the client aborting the request before it's completed.