Allow server-side to send formatted responses to console
Since we now report when the WhatsApp pipeline is executed from the console, a handler might want to send formatted markup to the console.

Reactions (since they contain a single character by default) are safe to render as markup (and could contain arbitrary additional markup). For text, we attempt to parse markup if text starts with `[` and ends with `]` which is a pretty good indicator.

We fallback to text in that case.