Add support for pluggable async message processing strategies
We were previously hardcoding the webhook > processing decoupling via the webjobs storage queue. This mechanism introduces some delay in production, so we now provide a pluggable mechanism with two providers:
- Queue processor: behaves like the previous implementation, unless explicitly overriden
- Event grid processor: delegates to an event grid topic and its accompanying event grid triggered function for processing.

Note that while queue processing Just Works without additional configuration, for Event Grid the user will have to create the event grid topic, configure it in the function app (topic url and access key) and the event subscription pointing back to the function.