Add typing indicator support during webhook or process
We now default to marking read and send typing indicator as soon as the webhook receives the message. This reduces user anxiety so we make that the default.

It can be configured by setting `WhatsApp:TypingOnMessage` to false and (for example) set `WhatApp:TypingOnProcess` to true. Same as for making as read (supported previously too).