Provide seamless continuity between CLI and WhatsApp
When not running in production, the new `UseConsole` can be used to support scenarios where the CLI falls short, such as sharing contacts, image or other media. In these cases, if the `UseConsole` is added after the `UseConversation` handler, you can get responses for messages in both WhatsApp as well as the CLI, provided a previous conversation in the time window came from the console (to establish its service url).