Model ConversationId as an extension property instead
The concept of a conversation id is not built-in WhatsApp, and therefore we should not force it on the basic API.

It's rather a concept we implement specifically as part of the conversation storage and handler, which are entirely optional.

With extension properties, we can make it seamless to use yet not tied to the core interfaces, since we support AdditionalProperties to support precisely this kind of extensibility scenarios.