Generalize interactive button to support lists selection too
Our previous InteractiveMessage assumed the payload to the webhook would always be a response to a button interactive message. The responses for list selection can be modeled similiarly since they contain id/title too.

So we make the interactive message more general by creating a new Selection record for both buttons and lists.