Allow inspecting the originating JSON for a message
Introduce non-serializable keys in the additional properties dictionary for messages (those that start with `__`) so that apps can leverage them for internal-only, transient data that doesn't need serialization at all.

Then right after deserializing a message, place the originating JSON as-is in that dictionary as `__json` and expose as internal property for debugging purposes. Note we explicitly don't make this part of the public API (for now?).