Allow invoking the processing behavior directly via HTTP POST
This can come in handy as a webhook-style mechanism that allows other arbitrary external queue/retry mechanisms to be implemented without requiring further changes to this library.

The only difference with the regular whatsapp_message endpoint is that the new whatsapp_process endpoint invokes the processing directly without doing any queuing whatesoever.