Improve URL paths for sub-commands
The main webhook route remains the same, but in order to avoid poluting the root path, we add / separator for cli, process and locally run eventgrid (via HTTP).

For now we keep the backwards-compatible /whatsappcli endpoint for older CLIs so they can continue to work but permanently redirecting to the new endpoint.