Speed-up initial webhook response times by going lazy init
Breaking change: we now *require* invoking UseWhatsApp on the functions app builder. We throw an exception if we detect it's misisng:

```
var builder = FunctionsApplication.CreateBuilder(args);
builder.ConfigureFunctionsWebApplication();
builder.UseWhatsApp();
```

When the user's handler pipeline was non-trivial to spin-up, webhook responses were delayed until the full DI graph for it was built.

We now instead make that dependency fully lazy so that the webhook can respond immediately and send the typing/read indicators as needed right-away.

Both pipeline runner and handler are changed to Func<T> accordingly, and we had to bring the function context accessor to properly retrieve the instances from the right container at run-time in the isolated worker model.