Allow dev processing event grid as webhook subscription
This allows a more realistic end to end testing leveraging the EventGrid settings if present and running as production even locally.

Configuration on the EventGrid Subscription should use Web Hook configuration and the /whatsapp/eventgrid endpoint that is exposed in this case (only for non-CI/Release builds).

We also default the initial value of the event grid options to an optional IOptions<EventGridOptions> if found, improving configurability without having to provide a configure lambda.